requirements.txt
numpy
pandas
redis
opencv-python-headless
scikit-learn
streamlit
streamlit-webrtc
insightface
onnxruntime
av
python-dotenv
